merge upstream

+++ app/extensions/extension_edit.php
@@ -35,7 +35,7 @@ else {
 }
 
 //detect billing app
-	$billing_app_exists = file_exists($_SERVER['DOCUMENT_ROOT'].PROJECT_PATH."/app/billing/app_config.php");
+	$billing_app_exists = file_exists($_SERVER["PROJECT_ROOT"]."/app/billing/app_config.php");
 	if ($billing_app_exists) {
 		require_once "app/billing/resources/functions/currency.php";
 		require_once "app/billing/resources/functions/rating.php";
@@ -240,7 +240,7 @@ else {
 							$sql_insert .= "domain_uuid, ";
 							$sql_insert .= "device_mac_address, ";
 							$sql_insert .= "device_template, ";
-							$sql_insert .= "device_provision_enable ";
+							$sql_insert .= "device_enabled ";
 							$sql_insert .= ") ";
 							$sql_insert .= "values ";
 							$sql_insert .= "(";
@@ -993,8 +993,7 @@ if (count($_POST) > 0 && strlen($_POST["persistformvar"]) == 0) {
 	echo "}\n";
 	echo "</script>";
 
-	echo "<form method='post' name='frm' action='' autocomplete='off'>\n";
-	echo "<input style='display:none;' type='password' name='autocomplete'>";
+	echo "<form method='post' name='frm' id='frm' action=''>\n";
 	echo "<table width='100%' border='0' cellpdding='0' cellspacing='0'>\n";
 	echo "<tr>\n";
 	if ($action == "add") {
@@ -1011,7 +1010,7 @@ if (count($_POST) > 0 && strlen($_POST["persistformvar"]) == 0) {
 	if ($action == "update") {
 		echo "	<input type='button' class='btn' alt='".$text['button-copy']."' onclick=\"copy_extension();\" value='".$text['button-copy']."'>\n";
 	}
-	echo "	<input type='submit' class='btn' value='".$text['button-save']."'>\n";
+	echo "	<input type='button' class='btn' value='".$text['button-save']."' onclick='submit_form();'>\n";
 	echo "	<br /><br />\n";
 	echo "</td>\n";
 	echo "</tr>\n";
@@ -1115,7 +1114,7 @@ if (count($_POST) > 0 && strlen($_POST["persistformvar"]) == 0) {
 				echo "		<tr>\n";
 				echo "			<td class='vtable'><a href='/core/users/usersupdate.php?id=".$field['user_uuid']."'>".$field['username']."</a></td>\n";
 				echo "			<td>\n";
-				echo "				<a href='#' onclick=\"if (confirm('".$text['confirm-delete']."')) { document.getElementById('delete_type').value = 'user'; document.getElementById('delete_uuid').value = '".$field['user_uuid']."'; document.forms.frm.submit(); }\" alt='".$text['button-delete']."'>$v_link_label_delete</a>\n";
+				echo "				<a href='#' onclick=\"if (confirm('".$text['confirm-delete']."')) { document.getElementById('delete_type').value = 'user'; document.getElementById('delete_uuid').value = '".$field['user_uuid']."'; submit_form(); }\" alt='".$text['button-delete']."'>$v_link_label_delete</a>\n";
 				//echo "				<a href='extension_edit.php?id=".$extension_uuid."&domain_uuid=".$_SESSION['domain_uuid']."&user_uuid=".$field['user_uuid']."&a=delete' alt='".$text['button-delete']."' onclick=\"return confirm('".$text['confirm-delete']."')\">$v_link_label_delete</a>\n";
 				echo "			</td>\n";
 				echo "		</tr>\n";
@@ -1130,8 +1129,8 @@ if (count($_POST) > 0 && strlen($_POST["persistformvar"]) == 0) {
 			echo "			<option value='".$field['user_uuid']."'>".$field['username']."</option>\n";
 		}
 		echo "			</select>";
-		echo "			<input type=\"submit\" class='btn' value=\"".$text['button-add']."\">\n";
-		
+		echo "			<input type='button' class='btn' value=\"".$text['button-add']."\" onclick='submit_form();'>\n";
+
 		echo "			<br>\n";
 		echo "			".$text['description-user_list']."\n";
 		echo "			<br />\n";
@@ -1190,7 +1189,7 @@ if (count($_POST) > 0 && strlen($_POST["persistformvar"]) == 0) {
 				echo "			<td class='vtable'>".$row['device_template']."&nbsp;</td>\n";
 				//echo "			<td class='vtable'>".$row['device_description']."&nbsp;</td>\n";
 				echo "			<td>\n";
-				echo "				<a href='#' onclick=\"if (confirm('".$text['confirm-delete']."')) { document.getElementById('delete_type').value = 'device_line'; document.getElementById('delete_uuid').value = '".$row['device_line_uuid']."'; document.forms.frm.submit(); }\" alt='".$text['button-delete']."'>$v_link_label_delete</a>\n";
+				echo "				<a href='#' onclick=\"if (confirm('".$text['confirm-delete']."')) { document.getElementById('delete_type').value = 'device_line'; document.getElementById('delete_uuid').value = '".$row['device_line_uuid']."'; submit_form(); }\" alt='".$text['button-delete']."'>$v_link_label_delete</a>\n";
 				echo "			</td>\n";
 				echo "		</tr>\n";
 			}
@@ -1301,7 +1300,7 @@ if (count($_POST) > 0 && strlen($_POST["persistformvar"]) == 0) {
 			echo "</select>\n";
 			echo "		</td>\n";
 			echo "		<td>\n";
-			echo "			<input type=\"submit\" class='btn' value=\"".$text['button-add']."\">\n";
+			echo "			<input type='button' class='btn' value=\"".$text['button-add']."\" onclick='submit_form();'>\n";
 			echo "		</td>\n";
 			echo "		</table>\n";
 			echo "		<br />\n";
@@ -1645,7 +1644,22 @@ if (count($_POST) > 0 && strlen($_POST["persistformvar"]) == 0) {
 		echo "    ".$text['label-toll_allow']."\n";
 		echo "</td>\n";
 		echo "<td class='vtable' align='left'>\n";
-		echo "    <input class='formfld' type='text' name='toll_allow' maxlength='255' value=\"$toll_allow\">\n";
+		if (is_array($_SESSION['toll allow']['name'])) {
+			echo "	<select class='formfld' name='toll_allow'>\n";
+			echo "		<option value=''></option>\n";
+			foreach ($_SESSION['toll allow']['name'] as $name) {
+				if ($_SESSION['call group']['name'] == $call_group) {
+					echo "		<option value='$name' selected='selected'>$name</option>\n";
+				}
+				else {
+					echo "		<option value='$name'>$name</option>\n";
+				}
+			}
+			echo "	</select>\n";
+		}
+		else {
+			echo "    <input class='formfld' type='text' name='toll_allow' maxlength='255' value=\"$toll_allow\">\n";
+		}
 		echo "<br />\n";
 		echo $text['description-toll_allow']."\n";
 		echo "</td>\n";
@@ -1668,7 +1682,21 @@ if (count($_POST) > 0 && strlen($_POST["persistformvar"]) == 0) {
 	echo "	".$text['label-call_group']."\n";
 	echo "</td>\n";
 	echo "<td class='vtable' align='left'>\n";
-	echo "	<input class='formfld' type='text' name='call_group' maxlength='255' value=\"$call_group\">\n";
+	if (is_array($_SESSION['call group']['name'])) {
+		echo "	<select class='formfld' name='call_group'>\n";
+		echo "		<option value=''></option>\n";
+		foreach ($_SESSION['call group']['name'] as $name) {
+			if ($_SESSION['call group']['name'] == $call_group) {
+				echo "		<option value='$name' selected='selected'>$name</option>\n";
+			}
+			else {
+				echo "		<option value='$name'>$name</option>\n";
+			}
+		}
+		echo "	</select>\n";
+	} else {
+		echo "	<input class='formfld' type='text' name='call_group' maxlength='255' value=\"$call_group\">\n";
+	}
 	echo "<br />\n";
 	echo $text['description-call_group']."\n";
 	echo "</td>\n";
@@ -1700,42 +1728,44 @@ if (count($_POST) > 0 && strlen($_POST["persistformvar"]) == 0) {
 		echo "</tr>\n";
 	}
 
-	echo "<tr>\n";
-	echo "<td class='vncell' valign='top' align='left' nowrap='nowrap'>\n";
-	echo "    ".$text['label-user_record']."\n";
-	echo "</td>\n";
-	echo "<td class='vtable' align='left'>\n";
-	echo "    <select class='formfld' name='user_record'>\n";
-	echo "    <option value=''>".$text['label-user_record_none']."</option>\n";
-	if ($user_record == "all") {
-		echo "    <option value='all' selected='selected'>".$text['label-user_record_all']."</option>\n";
+	if (permission_exists('extension_user_record')) {
+		echo "<tr>\n";
+		echo "<td class='vncell' valign='top' align='left' nowrap='nowrap'>\n";
+		echo "    ".$text['label-user_record']."\n";
+		echo "</td>\n";
+		echo "<td class='vtable' align='left'>\n";
+		echo "    <select class='formfld' name='user_record'>\n";
+		echo "    <option value=''>".$text['label-user_record_none']."</option>\n";
+		if ($user_record == "all") {
+			echo "    <option value='all' selected='selected'>".$text['label-user_record_all']."</option>\n";
+		}
+		else {
+			echo "    <option value='all'>".$text['label-user_record_all']."</option>\n";
+		}
+		if ($user_record == "local") {
+			echo "    <option value='local' selected='selected'>".$text['label-user_record_local']."</option>\n";
+		}
+		else {
+			echo "    <option value='local'>".$text['label-user_record_local']."</option>\n";
+		}
+		if ($user_record == "inbound") {
+			echo "    <option value='inbound' selected='selected'>".$text['label-user_record_inbound']."</option>\n";
+		}
+		else {
+			echo "    <option value='inbound'>".$text['label-user_record_inbound']."</option>\n";
+		}
+		if ($user_record == "outbound") {
+			echo "    <option value='outbound' selected='selected'>".$text['label-user_record_outbound']."</option>\n";
+		}
+		else {
+			echo "    <option value='outbound'>".$text['label-user_record_outbound']."</option>\n";
+		}
+		echo "    </select>\n";
+		echo "<br />\n";
+		echo $text['description-user_record']."\n";
+		echo "</td>\n";
+		echo "</tr>\n";
 	}
-	else {
-		echo "    <option value='all'>".$text['label-user_record_all']."</option>\n";
-	}
-	if ($user_record == "local") {
-		echo "    <option value='local' selected='selected'>".$text['label-user_record_local']."</option>\n";
-	}
-	else {
-		echo "    <option value='local'>".$text['label-user_record_local']."</option>\n";
-	}
-	if ($user_record == "inbound") {
-		echo "    <option value='inbound' selected='selected'>".$text['label-user_record_inbound']."</option>\n";
-	}
-	else {
-		echo "    <option value='inbound'>".$text['label-user_record_inbound']."</option>\n";
-	}
-	if ($user_record == "outbound") {
-		echo "    <option value='outbound' selected='selected'>".$text['label-user_record_outbound']."</option>\n";
-	}
-	else {
-		echo "    <option value='outbound'>".$text['label-user_record_outbound']."</option>\n";
-	}
-	echo "    </select>\n";
-	echo "<br />\n";
-	echo $text['description-user_record']."\n";
-	echo "</td>\n";
-	echo "</tr>\n";
 
 	if (is_dir($_SERVER["DOCUMENT_ROOT"].PROJECT_PATH.'/app/music_on_hold')) {
 		echo "<tr>\n";
@@ -1994,12 +2024,25 @@ if (count($_POST) > 0 && strlen($_POST["persistformvar"]) == 0) {
 		echo "		<input type='hidden' name='delete_uuid' id='delete_uuid' value=''>";
 	}
 	echo "			<br>";
-	echo "			<input type='submit' class='btn' value='".$text['button-save']."'>\n";
+	echo "			<input type='button' class='btn' value='".$text['button-save']."' onclick='submit_form();'>\n";
 	echo "		</td>\n";
 	echo "	</tr>";
 	echo "</table>";
 	echo "<br><br>";
 	echo "</form>";
+
+	echo "<script>\n";
+//capture enter key to submit form
+	echo "	$(window).keypress(function(event){\n";
+	echo "		if (event.which == 13) { submit_form(); }\n";
+	echo "	});\n";
+// convert password fields to
+	echo "	function submit_form() {\n";
+	echo "		$('input:password').css('visibility','hidden');\n";
+	echo "		$('input:password').attr({type:'text'});\n";
+	echo "		$('form#frm').submit();\n";
+	echo "	}\n";
+	echo "</script>\n";
 
 //include the footer
 	require_once "resources/footer.php";
